requirements.txt
google-genai==1.5.0
python-dotenv==1.0.0
colorama==0.4.6
requests==2.31.0
Pillow==10.2.0